+++ Auth/Exceptions/InvalidTokenException.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Dvsa\Contracts\Auth;
+namespace Dvsa\Contracts\Auth\Exceptions;
 
 /**
  * This exception will be thrown when trying to decode an invalid token.
  */
 class InvalidTokenException extends \Exception
 {
+    //
 }
